fix: #39568 - change preview.cjs to .jsx
- Muda a porta storybook 6006 para 9009 para evitar conflitos caso rode junto ao sistema perfarm;
- Foi necessário mudar o preview para .jsx pois estava resultando em conflitos durante o yarn storybook, o mesmo não acusou erro após mudança, nem durante o yarn build
;
