+++ .storybook/main.cjs
@@ -11,14 +11,4 @@ module.exports = {
     };
     return config;
   },
-  previewHead: (head) => `
-    ${head}
-    <style>
-      *,
-      ::after,
-      ::before {
-        box-sizing: border-box;
-      }
-    </style>
-  `,
 };
