+++ src/forms/MachineForm/stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { globalStyles } from '~/config/globalStyles';
 import { MachineForm } from './index';
 
 export default {
@@ -8,11 +7,7 @@ export default {
   component: MachineForm,
 } as ComponentMeta<typeof MachineForm>;
 
-const Template: ComponentStory<typeof MachineForm> = (args) => {
-  globalStyles();
-
-  return <MachineForm {...args} />;
-};
+const Template: ComponentStory<typeof MachineForm> = (args) => <MachineForm {...args} />;
 
 export const Primary = Template.bind({});
